Added Search Bar Scaffolding
